[metrics] move pusher mod to separate file
